Stip attrs from test_ids before updating times db
This address the issue where timing data is thrown away from previous
runs if the attrs set on a test change. This was caused by the times
db using the test_id with the attrs as the key in the db. This commit
fixes this by stripping attrs before storing them in the db and then
stripping attrs from the test_ids in the scheduler before looking up
timing data.

Fixes-bug: #1416512